API: move the rest of the recovery methods to recover.py
